Optimization: change `smearingPeriod` type from uint256 to uint24

+++ src/EulerEarnFactory.sol
@@ -42,7 +42,7 @@ contract EulerEarnFactory is IEulerEarnFactory {
         string memory _name,
         string memory _symbol,
         uint256 _initialCashAllocationPoints,
-        uint256 _smearingPeriod
+        uint24 _smearingPeriod
     ) external returns (address) {
         address eulerEulerEarnVault = Clones.clone(eulerEarnImpl);
